Improve structure of react app
