+++ .eslintrc
@@ -18,6 +18,7 @@
     "generator-star-spacing": "off",
     "import/no-unresolved": "error",
     "import/no-extraneous-dependencies": "off",
+    "jsx-a11y/href-no-hash": "off",
     "no-console": "off",
     "no-use-before-define": "off",
     "no-underscore-dangle": "off",
