+++ app/app.global.scss
@@ -1,5 +1,5 @@
 @import './font/fontello.css';
-@import './components/variables.scss';
+@import './components/common/variables.scss';
 
 *,
 *:before,
