+++ app/components/Main/Home.scss
