+++ app/components/Main/LoadingFail.scss
