+++ app/components/Main/UpdateBar.scss
@@ -1,4 +1,4 @@
-@import './variables.scss';
+@import '../common/variables.scss';
 $base-color: #444444;
 
 .updateBar {
