+++ app/components/Main/WebView.scss
@@ -1,4 +1,4 @@
-@import './variables.scss';
+@import '../common/variables.scss';
 
 .webview {
     position: absolute;
