+++ app/components/common/variables.scss
