--- app/containers/AboutPage.js
@@ -1,16 +1,0 @@
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
-import About from '../components/About';
-import * as CounterActions from '../actions/counter';
-
-function mapStateToProps(state) {
-  return {
-    counter: state.counter
-  };
-}
-
-function mapDispatchToProps(dispatch) {
-  return bindActionCreators(CounterActions, dispatch);
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(About);
