+++ app/containers/AboutPage.jsx
@@ -1,0 +1,9 @@
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux';
+import About from '../components/About';
+
+function mapStateToProps(state) {
+  return {};
+}
+
+export default connect(mapStateToProps)(About);
