+++ app/containers/HomePage.jsx
@@ -1,12 +1,11 @@
 // @flow
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import Home from '../components/Home';
+import Main from '../components/Main';
 import * as ServerActions from '../actions/server';
 
 function mapStateToProps(state) {
   return {
-    counter: state.counter,
     servers: state.servers,
     active: state.activeServer,
     sidebarStatus: state.sidebarStatus,
@@ -18,4 +17,4 @@ function mapDispatchToProps(dispatch) {
   return bindActionCreators(ServerActions, dispatch);
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Home);
+export default connect(mapStateToProps, mapDispatchToProps)(Main);
