+++ app/routes.jsx
@@ -3,13 +3,11 @@ import React from 'react';
 import { Switch, Route } from 'react-router';
 import App from './containers/App';
 import HomePage from './containers/HomePage';
-import CounterPage from './containers/CounterPage';
 import AboutPage from './containers/AboutPage';
 
 export default () => (
   <App>
     <Switch>
-      <Route path="/counter" component={CounterPage} />
       <Route path="/about" component={AboutPage} />
       <Route path="/" component={HomePage} />
     </Switch>
