+++ app/store/configureStore.prod.js
@@ -5,11 +5,10 @@ import { createHashHistory } from 'history';
 import { electronEnhancer } from 'redux-electron-store';
 import { routerMiddleware } from 'react-router-redux';
 import rootReducer from '../reducers';
-import type { counterStateType } from '../reducers/counter';
 
 const history = createHashHistory();
 
-function configureStore(initialState?: counterStateType) {
+function configureStore(initialState) {
   const router = routerMiddleware(history);
   const enhancer = compose(
     applyMiddleware(thunk, router),
