+++ package.json
@@ -25,7 +25,7 @@
     "package-all": "npm run build && build -mwl",
     "package-linux": "npm run build && build --linux",
     "package-win": "npm run build && build --win --x64",
-    "postinstall": "concurrently \"npm run flow-typed\" \"electron-builder install-app-deps\" \"node node_modules/fbjs-scripts/node/check-dev-engines.js package.json\"",
+    "postinstall": "concurrently \"npm run flow-typed\" \"npm run build-dll\" \"electron-builder install-app-deps\" \"node node_modules/fbjs-scripts/node/check-dev-engines.js package.json\"",
     "release": "npm run build && build",
     "start": "cross-env START_HOT=1 npm run start-renderer-dev",
     "prestart:prod": "npm run build",
@@ -120,7 +120,7 @@
   "devDependencies": {
     "babel-core": "^6.26.0",
     "babel-eslint": "^7.2.3",
-    "babel-jest": "^20.0.3",
+    "babel-jest": "^21.0.0",
     "babel-loader": "^7.1.2",
     "babel-minify-webpack-plugin": "^0.2.0",
     "babel-plugin-add-module-exports": "^0.2.1",
@@ -151,8 +151,8 @@
     "eslint-plugin-flowtype": "^2.35.1",
     "eslint-plugin-flowtype-errors": "^3.3.1",
     "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jest": "^20.0.3",
-    "eslint-plugin-jsx-a11y": "6.0.2",
+    "eslint-plugin-jest": "^21.0.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
     "eslint-plugin-promise": "^3.5.0",
     "eslint-plugin-react": "^7.3.0",
     "extract-text-webpack-plugin": "^3.0.0",
@@ -163,7 +163,7 @@
     "flow-typed": "^2.1.5",
     "html-webpack-plugin": "^2.30.1",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^20.0.4",
+    "jest": "^21.0.1",
     "jsdom": "^11.2.0",
     "minimist": "^1.2.0",
     "node-sass": "^4.5.3",
